+++ packages/core/src/main.ts
@@ -1,72 +1,64 @@
-import {
-  ConnectedStarknetWindowObject,
-  StarknetWindowObject,
-} from "./StarknetWindowObject"
-import discovery, { WalletProvider } from "./discovery"
-import { IStorageWrapper, LocalStorageWrapper } from "./localStorageStore"
+import discovery, { type WalletProvider } from "./discovery"
+import { LocalStorageWrapper } from "./localStorageStore"
+import type { GetStarknetOptions, GetStarknetResult } from "./types"
 import { pipe, ssrSafeWindow } from "./utils"
-import { FilterList, filterBy, filterByPreAuthorized } from "./wallet/filter"
-import { isWalletObj } from "./wallet/isWalletObject"
-import { injectMetamaskBridge } from "./wallet/metamaskBridge"
+import { filterBy, filterByAuthorized } from "./wallet/filter"
+import {
+  isFullWallet,
+  isVirtualWallet,
+  isWalletObject,
+} from "./wallet/isWalletObject"
 import { scanObjectForWallets } from "./wallet/scan"
-import { Sort, sortBy } from "./wallet/sort"
+import { sortBy } from "./wallet/sort"
+import {
+  initiateVirtualWallets,
+  resolveVirtualWallet,
+  virtualWallets,
+} from "./wallet/virtualWallets"
+import { Permission, type StarknetWindowObject } from "@starknet-io/types-js"
 
 export type {
-  AccountChangeEventHandler,
-  AddStarknetChainParameters,
-  ConnectedStarknetWindowObject,
-  NetworkChangeEventHandler,
-  RpcMessage,
   StarknetWindowObject,
-  SwitchStarknetChainParameter,
-  WalletEvents,
+  AddDeclareTransactionParameters,
+  AddDeclareTransactionResult,
+  AddInvokeTransactionParameters,
+  AddInvokeTransactionResult,
+  AddStarknetChainParameters,
+  RequestAccountsParameters,
+  SwitchStarknetChainParameters,
+  AccountDeploymentData,
   WatchAssetParameters,
-  DisconnectedStarknetWindowObject,
-  IStarknetWindowObject,
-} from "./StarknetWindowObject"
-export type {
-  WalletProvider,
-  BrowserStoreVersion,
-  OperatingSystemStoreVersion,
-} from "./discovery"
+  TypedData,
+  RequestFn,
+  RpcMessage,
+  IsParamsOptional,
+  RpcTypeToMessageMap,
+  RequestFnCall,
+  AccountChangeEventHandler,
+  NetworkChangeEventHandler,
+  WalletEventHandlers,
+  WalletEvents,
+} from "@starknet-io/types-js"
 
-export interface GetStarknetOptions {
-  windowObject: Record<string, any>
-  isWalletObject: (wallet: any) => boolean
-  storageFactoryImplementation: (name: string) => IStorageWrapper
-}
+export { scanObjectForWallets } from "./wallet/scan"
+export { isWalletObject } from "./wallet/isWalletObject"
+
+export type {
+  BrowserStoreVersion,
+  DisconnectOptions,
+  GetStarknetOptions,
+  GetStarknetResult,
+  GetWalletOptions,
+  OperatingSystemStoreVersion,
+  WalletProvider,
+} from "./types"
+
+export { ssrSafeWindow } from "./utils"
 
 const defaultOptions: GetStarknetOptions = {
   windowObject: ssrSafeWindow ?? {},
-  isWalletObject: isWalletObj,
+  isWalletObject,
   storageFactoryImplementation: (name: string) => new LocalStorageWrapper(name),
-}
-
-export interface GetWalletOptions {
-  sort?: Sort
-  include?: FilterList
-  exclude?: FilterList
-}
-
-export interface DisconnectOptions {
-  clearLastWallet?: boolean
-}
-interface GetStarknetResult {
-  getAvailableWallets: (
-    options?: GetWalletOptions,
-  ) => Promise<StarknetWindowObject[]> // Returns all wallets available in the window object
-  getPreAuthorizedWallets: (
-    options?: GetWalletOptions,
-  ) => Promise<StarknetWindowObject[]> // Returns only preauthorized wallets available in the window object
-  getDiscoveryWallets: (options?: GetWalletOptions) => Promise<WalletProvider[]> // Returns all wallets in existence (from discovery file)
-  getLastConnectedWallet: () => Promise<StarknetWindowObject | null | undefined> // Returns the last wallet connected when it's still connected
-  enable: (
-    wallet: StarknetWindowObject,
-    options?: {
-      starknetVersion?: "v4" | "v5"
-    },
-  ) => Promise<ConnectedStarknetWindowObject> // Connects to a wallet
-  disconnect: (options?: DisconnectOptions) => Promise<void> // Disconnects from a wallet
 }
 
 export function getStarknet(
@@ -78,7 +70,7 @@ export function getStarknet(
   }
   const lastConnectedStore = storageFactoryImplementation("gsw-last")
 
-  injectMetamaskBridge(windowObject)
+  initiateVirtualWallets(windowObject)
 
   return {
     getAvailableWallets: async (options = {}) => {
@@ -91,13 +83,13 @@ export function getStarknet(
         (_) => sortBy(_, options.sort),
       )(availableWallets)
     },
-    getPreAuthorizedWallets: async (options = {}) => {
+    getAuthorizedWallets: async (options = {}) => {
       const availableWallets = scanObjectForWallets(
         windowObject,
         isWalletObject,
       )
       return pipe<StarknetWindowObject[]>(
-        (_) => filterByPreAuthorized(_),
+        (_) => filterByAuthorized(_),
         (_) => filterBy(_, options),
         (_) => sortBy(_, options.sort),
       )(availableWallets)
@@ -114,20 +106,62 @@ export function getStarknet(
       const lastConnectedWallet = allWallets.find(
         (w) => w.id === lastConnectedWalletId,
       )
-      const [firstPreAuthorizedWallet] = await filterByPreAuthorized(
+      const [firstAuthorizedWallet] = await filterByAuthorized(
         lastConnectedWallet ? [lastConnectedWallet] : [],
       )
 
-      if (!firstPreAuthorizedWallet) {
+      if (!firstAuthorizedWallet) {
         lastConnectedStore.delete()
         return null
       }
 
-      return firstPreAuthorizedWallet
+      return firstAuthorizedWallet
     },
-    enable: async (wallet, options) => {
-      await wallet.enable(options ?? { starknetVersion: "v5" })
-      if (!wallet.isConnected) {
+    discoverVirtualWallets: async (
+      walletNamesOrIds: string[] = [],
+    ): Promise<void> => {
+      const walletNamesOrIdsSet = new Set(walletNamesOrIds)
+
+      const virtualWalletToDiscover =
+        walletNamesOrIdsSet.size > 0
+          ? virtualWallets.filter(
+              (virtualWallet) =>
+                walletNamesOrIdsSet.has(virtualWallet.name) ||
+                walletNamesOrIdsSet.has(virtualWallet.id),
+            )
+          : virtualWallets
+
+      await Promise.all(
+        virtualWalletToDiscover.map(async (virtualWallet) => {
+          const hasSupport = await virtualWallet.hasSupport(windowObject)
+          if (hasSupport) {
+            windowObject[virtualWallet.windowKey] = virtualWallet
+          }
+        }),
+      )
+    },
+    enable: async (inputWallet, options) => {
+      let wallet: StarknetWindowObject
+      if (isVirtualWallet(inputWallet)) {
+        wallet = await resolveVirtualWallet(windowObject, inputWallet)
+      } else if (isFullWallet(inputWallet)) {
+        wallet = inputWallet
+      } else {
+        throw new Error("Invalid wallet object")
+      }
+
+      await wallet.request({
+        type: "wallet_requestAccounts",
+        params: {
+          silent_mode: options?.silent_mode,
+        },
+      })
+
+      // check for permissions
+      const permissions: Permission[] = await wallet.request({
+        type: "wallet_getPermissions",
+      })
+      if (!permissions?.includes(Permission.ACCOUNTS)) {
         throw new Error("Failed to connect to wallet")
       }
       lastConnectedStore.set(wallet.id)
@@ -140,7 +174,5 @@ export function getStarknet(
     },
   }
 }
-
-export { ssrSafeWindow } from "./utils"
 
 export default getStarknet()