+++ example/CHANGELOG.md
@@ -1,4 +1,49 @@
-# get-starknet-example
+# @starknet-io/get-starknet-example
+
+## 4.0.6
+
+### Patch Changes
+
+- @starknet-io/get-starknet@4.0.6
+
+## 4.0.5
+
+### Patch Changes
+
+- @starknet-io/get-starknet@4.0.5
+
+## 4.0.4
+
+### Patch Changes
+
+- @starknet-io/get-starknet@4.0.4
+
+## 4.0.3
+
+### Patch Changes
+
+- @starknet-io/get-starknet@4.0.3
+
+## 4.0.2
+
+### Patch Changes
+
+- @starknet-io/get-starknet@4.0.2
+
+## 4.0.1
+
+### Patch Changes
+
+- 0cd43d0: added mobile os store links support
+- Updated dependencies [0cd43d0]
+  - @starknet-io/get-starknet@4.0.1
+
+## 2.1.3-next.0
+
+### Patch Changes
+
+- Updated dependencies [4a477fd]
+  - get-starknet@4.0.0-next.0
 
 ## 2.1.4